refactor: move all composable functions to dedicated folder
