+++ src/components/Articles/Preview.vue
@@ -51,7 +51,7 @@
 
 <script lang="ts">
 import { computed, defineComponent, PropType } from 'vue'
-import { useFavoriteArticle } from '../../services/article/favoriteArticle'
+import { useFavoriteArticle } from '../../composable/useFavoriteArticle'
 
 export default defineComponent({
   name: 'ArticlePreview',
