+++ src/services/profile/followProfile.ts
@@ -1,11 +1,7 @@
-import type { ComputedRef } from 'vue'
-import { redirect } from '../../router'
-
 import type { AuthorizationError } from '../../types/error'
 
 import { request } from '../index'
 
-import createAsyncProcess from '../../utils/create-async-process'
 import { mapAuthorizationResponse } from '../../utils/map-checkable-response'
 import { Either, fail, success } from '../../utils/either'
 
@@ -23,32 +19,4 @@ export async function deleteFollowProfile (username: string): Promise<Either<Aut
 
   if (result2.isOk()) return success(result2.value.profile)
   return fail(result2.value)
-}
-
-interface UseFollowProps {
-  username: ComputedRef<string>
-  following: ComputedRef<boolean>
-  onUpdate: (profile: Profile) => void
-}
-
-export function useFollow ({ username, following, onUpdate }: UseFollowProps) {
-  async function toggleFollow () {
-    let response: Either<AuthorizationError, Profile>
-
-    if (following.value === true) {
-      response = await deleteFollowProfile(username.value)
-    } else {
-      response = await postFollowProfile(username.value)
-    }
-
-    if (response.isOk()) onUpdate(response.value)
-    else redirect('login')
-  }
-
-  const { active, run } = createAsyncProcess(toggleFollow)
-
-  return {
-    followProcessGoing: active,
-    toggleFollow: run,
-  }
 }
