+++ src/services/profile/getProfile.ts
@@ -1,28 +1,5 @@
-import { reactive, watchEffect } from 'vue'
 import { request } from '../index'
 
 export async function getProfile (username: string) {
   return request.get<ProfileResponse>(`/profiles/${username}`).then(res => res.profile)
-}
-
-export function useProfile (username: string) {
-  const profile = reactive<Profile>({} as Profile)
-
-  async function fetchProfile () {
-    const profileData = await getProfile(username)
-    updateProfile(profileData)
-  }
-
-  async function updateProfile (profileData: Profile) {
-    Object.assign(profile, profileData)
-  }
-
-  watchEffect(() => {
-    fetchProfile()
-  })
-
-  return {
-    profile,
-    updateProfile,
-  }
 }
